Introduce Scheduler termination
This is a signal-based mechanism that integrates with both `EventLoop`
and `Fiber`. It is expected that the application installs a signal
handler that eventually calls `Scheduler::terminate`. The signal sent
also interrupts the `EventLoop` if it's waiting on the kernel for
events.  Any blocked fibers will be interrupted, and it is the
responsibility of the async primitives to handle the interruption event.

The interruption mechanism helped uncover a leak in `Scheduler`. I
noticed that not all `Fiber` instances were cleaned up (destructor
wasn't being called) and it turns out that ASan support was incomplete:
there is a need to call fiber-specific hooks that keep track of the
stack across context switches. The erroneous code generates the
following leak report when running the example async_echo_server.

```
async_echo_server(5531,0x200658240) malloc: nano zone abandoned due to inability to reserve vm space.
Echo server launched on 127.0.0.1:8080
^Ccould not accept connection - error: 4

=================================================================
==5531==ERROR: LeakSanitizer: detected memory leaks

Indirect leak of 33064 byte(s) in 1 object(s) allocated from:
    #0 0x0001055ba370 in _Znwm+0x6c (libclang_rt.asan_osx_dynamic.dylib:arm64+0x62370)
    #1 0x000104ec2aec in axle::Scheduler::Scheduler(std::__1::shared_ptr<axle::EventLoop>) scheduler.cpp:49
    #2 0x000104ec119c in axle::Scheduler::init() scheduler.cpp:19
    #3 0x000104eafb60 in main main.cpp:94
    #4 0x0001969f8270  (<unknown module>)

Indirect leak of 33064 byte(s) in 1 object(s) allocated from:
    #0 0x0001055ba370 in _Znwm+0x6c (libclang_rt.asan_osx_dynamic.dylib:arm64+0x62370)
    #1 0x000104ec17dc in axle::Scheduler::do_schedule(std::__1::function<void ()>) scheduler.cpp:55
    #2 0x000104ec147c in axle::Scheduler::schedule(std::__1::function<void ()>) scheduler.cpp:28
    #3 0x000104eafb8c in main main.cpp:95
    #4 0x0001969f8270  (<unknown module>)

Indirect leak of 48 byte(s) in 1 object(s) allocated from:
    #0 0x0001055ba370 in _Znwm+0x6c (libclang_rt.asan_osx_dynamic.dylib:arm64+0x62370)
    #1 0x000104ec18b4 in axle::Scheduler::do_schedule(std::__1::function<void ()>) scheduler.cpp:55
    #2 0x000104ec147c in axle::Scheduler::schedule(std::__1::function<void ()>) scheduler.cpp:28
    #3 0x000104eafb8c in main main.cpp:95
    #4 0x0001969f8270  (<unknown module>)

-----------------------------------------------------
Suppressions used:
  count      bytes template
      3        120 *_fetchInitializingClassList
-----------------------------------------------------
```

PR: https://github.com/whalbawi/axle/pull/18